refactor proposal generation routes
